% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlickrAPIRequest.R
\name{FlickAPIRequest}
\alias{FlickAPIRequest}
\title{Request data from the Flickr API}
\usage{
FlickAPIRequest(method = NULL, api_key = NULL, ...)
}
\arguments{
\item{method}{Flickr API method to use for request.}

\item{api_key}{Flickr API key. If api_key is `NULL`, the function uses
[getFlickrAPIKey()] to use the environment variable "FLICKR_API_KEY" as the
key.}

\item{...}{Additional parameters passed to `httr2::req_url_query`}
}
\description{
Request data from the Flickr API using the provided method, API key, and any
additional values passed to `httr2::req_url_query`.
}
