% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExif.R
\name{getExif}
\alias{getExif}
\title{Exif Data}
\usage{
getExif(api_key, photo_id)
}
\arguments{
\item{api_key}{Your API key generated by your Flickr developper account}

\item{photo_id}{Specific characters/numbers of photo id of which you would like to get information}
}
\value{
This function returns a dataframe of Exif information of given photograph
}
\description{
Exif data of a specific photograph
}
\examples{
\dontrun{
  getExif(api_key="XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX", photo_id="45961963324")
}

}
\author{
Koki Ando <koki.25.ando@gmail.com>
}
