% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPhotoInfo.R
\name{getPhotoInfo}
\alias{getPhotoInfo}
\title{Download Flickr Photo Data}
\usage{
getPhotoInfo(api_key, photo_id, output = c("Location", "Date", "URL",
  "Tags"))
}
\arguments{
\item{api_key}{Your API key generated by your Flickr developper account}

\item{photo_id}{Specific characters/numbers of photo id of which you would like to get information}

\item{output}{Output data type. It depends on what you are interested in. You have to choose from Location, Date, URL or Tags.}
}
\value{
This function returns data of specific photo's information.
}
\description{
R access to photo information of photos posted on flickr via flickr API interface
}
\examples{
\dontrun{
  getPhotoInfo(api_key = "XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX", photo_id="30484882493", 
               output="location")
}

}
\author{
Koki Ando <koki.25.ando@gmail.com>
}
