% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPhotos.R
\name{getPhotos}
\alias{getPhotos}
\title{Photos from the given user's photostream}
\usage{
getPhotos(api_key, user_id)
}
\arguments{
\item{api_key}{Your API application key}

\item{user_id}{The NSID of the user who's phoyos to return. A value of me return the calling user's photos}
}
\value{
This function return \code{data.frame} including columns:
\itemize{
 \item id
 \item owner
 \item secret
 \item server
 \item farm
 \item title
 \item ispublic
 \item isfriend
 \item isfamily
}
}
\description{
Returns photos from the given user's photostream. Only phots visible to the calling user will be returned.
}
\examples{
\dontrun{
  getPhotos(api_key="XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX", user_id="141696738@N08")
}

}
\author{
Koki Ando <koki.25.ando@gmail.com>
}
