% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPhotoSearch.R
\name{getPhotoSearch}
\alias{getPhotoSearch}
\title{Search for photos on Flickr by user id, tags, license, or bounding box}
\usage{
getPhotoSearch(
  api_key,
  user_id,
  tags,
  licence_id,
  bbox,
  extras,
  per_page = 100,
  page = 1
)
}
\arguments{
\item{api_key}{Required. Your API application key}

\item{user_id}{The NSID of the user who's photo to search. If this parameter isn't passed then everybody's public photos will be searched.}

\item{tags}{List of tags you wish to search for e.g. c("cats, "dogs)}

\item{bbox}{A object of class bbox or a comma-delimited list of 4 values defining the Bounding Box of the area that will be searched. The 4 values represent the bottom-left corner of the box and the top-right corner, minimum_longitude, minimum_latitude, maximum_longitude, maximum_latitude.}

\item{extras}{A vector of extra information to fetch for each returned record. Currently supported fields are: c("description", "license", "date_upload", "date_taken", "owner_name", "icon_server", "original_format", "last_update", "geo", "tags", "machine_tags", "o_dims", "views", "media", "path_alias", "url_sq", "url_t", "url_s", "url_q", "url_m", "url_n", "url_z", "url_c", "url_l", "url_o")}

\item{per_page}{Number specifying how many results per page to return. Default 100 results per page. Maximum of 250 if bbox provided or 500 otherwise.}

\item{page}{Number specifying which search results page to return. Default is page 1 of results returned.}

\item{license_id}{The license id for photos (for possible values see the Flickr API method flickr.photos.licenses.getInfo).}
}
\value{
This function returns data of specific photos matching search parameters.
}
\description{
R access to flickr search api
}
\examples{
\dontrun{
  getPhotoSearch(api_key = "XXXXXXXXXX",
               tags= "cats")
}
}
