% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlickrAPI-package.R
\docType{package}
\name{FlickrAPI-package}
\alias{FlickrAPI}
\alias{FlickrAPI-package}
\title{FlickrAPI: Access to Flickr API via R}
\description{
An R package that provides an interface to the Flickr API <https://www.flickr.com/services/api/>,
which allows R users to download publicly available data on Flickr feed such as location, exif,
list of photos posted by specific user.
}
\seealso{
\url{https://www.flickr.com/services/api/}
}
\author{
Koki Ando \email{koki.25.ando@gmail.com}
}
\keyword{internal}
