% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlickrAPIRequest.R
\name{FlickrAPIRequest}
\alias{FlickrAPIRequest}
\title{Request data from the Flickr API}
\usage{
FlickrAPIRequest(
  method = NULL,
  api_key = NULL,
  format = "json",
  simplifyVector = TRUE,
  check_type = FALSE,
  ...
)
}
\arguments{
\item{method}{Flickr API method to use for request.}

\item{api_key}{Flickr API key. If api_key is \code{NULL}, the function uses
\code{\link[=getFlickrAPIKey]{getFlickrAPIKey()}} to use the environment variable "FLICKR_API_KEY" as the
key.}

\item{format}{Format parameter passed to \link[httr2:req_url]{httr2::req_url_query}}

\item{simplifyVector}{Default to \code{TRUE}, passed to \link[httr2:resp_body_raw]{httr2::resp_body_json}}

\item{check_type}{Default to \code{FALSE}, passed to \link[httr2:resp_body_raw]{httr2::resp_body_json}}

\item{...}{Additional parameters passed to \link[httr2:req_url]{httr2::req_url_query}}
}
\description{
Request data from the Flickr API using the provided method, API key, and any
additional values passed to \link[httr2:req_url]{httr2::req_url_query}.
}
