% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExif.R
\name{getExif}
\alias{getExif}
\alias{get_exif}
\title{Get EXIF data for a Flickr photo.}
\usage{
getExif(api_key = NULL, photo_id = NULL)

get_exif(api_key = NULL, photo_id = NULL)
}
\arguments{
\item{api_key}{Flickr API key. If api_key is \code{NULL}, the function uses
\code{\link[=getFlickrAPIKey]{getFlickrAPIKey()}} to use the environment variable "FLICKR_API_KEY" as the
key.}

\item{photo_id}{The id of the photo to fetch information for}
}
\value{
This function returns a data frame of EXIF information of given
photograph
}
\description{
Return a data of EXIF data for a given photo. The calling user must have
permission to view the photo.
}
\examples{
\dontrun{
getExif(
api_key = get_flickr_api_key(),
photo_id = "45961963324"
)
}

}
