% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExif.R
\name{getExif}
\alias{getExif}
\alias{get_exif}
\title{Returns a data of EXIF data for a given photo.}
\usage{
getExif(api_key = NULL, photo_id)
}
\arguments{
\item{api_key}{Your API application key}

\item{photo_id}{The id of the photo to fetch information for}
}
\value{
This function returns a data frame of EXIF information of given
  photograph
}
\description{
Return a data of EXIF data for a given photo. The calling user must have
permission to view the photo.
}
\examples{
\dontrun{
getExif(api_key = "XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX", photo_id = "45961963324")
}

}
