% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHotTags.R
\name{getHotTags}
\alias{getHotTags}
\title{Returns a list of hot tags for the given period.}
\usage{
getHotTags(api_key, period = "day", count = 20)
}
\arguments{
\item{api_key}{Your API application key}

\item{period}{The period for which to fetch hot tags. Valid values are day and week. Defaults to day}

\item{count}{The number of tags to return. Defaults to 20. Maximum allowed value is 200.}
}
\value{
This function a data frame of hot tags for the given period
}
\description{
Returns a list of hot tags for the given period.
}
\examples{
\dontrun{
 getHotTags(api_key = "********************************", 
            period = "day", count=20)
}

}
\author{
Koki Ando
}
