% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPhotoInfo.R
\name{getPhotoInfo}
\alias{getPhotoInfo}
\title{Download Flickr Photo Data}
\usage{
getPhotoInfo(api_key, photo_id, output = c("Location", "Date", "URL", "Tags"))
}
\arguments{
\item{api_key}{Your API application key}

\item{photo_id}{The id of the photo to get information for}

\item{output}{Output data type. It depends on what you are interested in. You have to choose from Location, Date, URL or Tags.}
}
\value{
This function returns data of specific photo's information.
}
\description{
R access to photo information of photos posted on flickr via flickr API interface
}
\examples{
\dontrun{
  getPhotoInfo(api_key = "XXXXXXXXXX",
               photo_id="30484882493",
               output="location")
}

}
\seealso{
\url{https://www.flickr.com/services/api/flickr.photos.getInfo.html}
}
\author{
Koki Ando <koki.25.ando@gmail.com>
}
