% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPhotos.R
\name{getPhotos}
\alias{getPhotos}
\title{Photo's information}
\usage{
getPhotos(api_key, user_id)
}
\arguments{
\item{api_key}{Your API key generated by your Flickr developper account}

\item{user_id}{Specific string of user id of which you would like to get information}
}
\value{
This function return \code{data.frame} including columns:
\itemize{
 \item id
 \item owner
 \item secret
 \item server
 \item farm
 \item title
 \item ispublic
 \item isfriend
 \item isfamily
}
}
\description{
Data of photos posted by specific Flickr account.
}
\examples{
\dontrun{
  getPhotos(api_key="XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX", user_id="141696738@N08")
}

}
\author{
Koki Ando <koki.25.ando@gmail.com>
}
