% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPhotos.R
\name{getPhotos}
\alias{getPhotos}
\alias{get_photos}
\title{Get photos from a Flickr user photo stream.}
\usage{
getPhotos(api_key = NULL, user_id)

get_photos(api_key = NULL, user_id)
}
\arguments{
\item{api_key}{Flickr API key. If api_key is `NULL`, the function uses
[getFlickrAPIKey()] to use the environment variable "FLICKR_API_KEY" as the
key.}

\item{user_id}{The NSID of the user who's photos to return. A value of "me"
return the calling user's photos.}
}
\value{
This function returns a \code{data.frame} including columns:
\itemize{
 \item id
 \item owner
 \item secret
 \item server
 \item farm
 \item title
 \item ispublic
 \item isfriend
 \item isfamily
}
}
\description{
Returns photos from the given user's photo stream. Only photos visible to the
calling user will be returned.
}
\examples{
\dontrun{
getPhotos(api_key = "XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX", user_id = "141696738@N08")
}
}
