% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPhotos.R
\name{getPhotos}
\alias{getPhotos}
\alias{get_photos}
\title{Get photos from a Flickr user photo stream.}
\usage{
getPhotos(user_id = NULL, extras = NULL, api_key = NULL, ...)

get_photos(user_id = NULL, extras = NULL, api_key = NULL, ...)
}
\arguments{
\item{user_id}{The NSID of the user who's photos to return. A value of "me"
return the calling user's photos.}

\item{extras}{A vector of extra information to fetch for each returned
record. Currently supported fields are: c("description", "license",
"date_upload", "date_taken", "owner_name", "icon_server",
"original_format", "last_update", "geo", "tags", "machine_tags", "o_dims",
"views", "media", "path_alias", "url_sq", "url_t", "url_s", "url_q",
"url_m", "url_n", "url_z", "url_c", "url_l", "url_o")}

\item{api_key}{Flickr API key. If api_key is `NULL`, the function uses
[getFlickrAPIKey()] to use the environment variable "FLICKR_API_KEY" as the
key.}

\item{...}{Additional parameters passed to [getPhotoSearch]}
}
\value{
This function returns a \code{data.frame} including columns:
\itemize{
 \item id
 \item owner
 \item secret
 \item server
 \item farm
 \item title
 \item ispublic
 \item isfriend
 \item isfamily
}
}
\description{
Returns photos from the given user's photo stream. Only photos visible to the
calling user will be returned.
}
\examples{
\dontrun{
getPhotos(api_key = "XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX", user_id = "141696738@N08")
}
}
