% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPhotoSearch.R
\name{getPhotoSearch}
\alias{getPhotoSearch}
\title{Search for photos on Flickr by tags}
\usage{
getPhotoSearch(api_key, tags, page = 1, per_page = 500)
}
\arguments{
\item{api_key}{Your API application key}

\item{tags}{List of tags you wish to search for e.g. c("cats, "dogs)}

\item{page}{Number specifying which search results page to return. Default is page 1 of results returned.}

\item{per_page}{Number specifying how many results per page to return. Default 500 results per page.}
}
\value{
This function returns data of specific photos matching search tags.
}
\description{
R access to flickr tag search api
}
\examples{
\dontrun{
  getPhotoSearch(api_key = "XXXXXXXXXX",
               tags= "cats")
}
}
