% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlickrAPIKey.R
\name{setFlickrAPIKey}
\alias{setFlickrAPIKey}
\alias{set_flickr_api_key}
\alias{getFlickrAPIKey}
\alias{get_flickr_api_key}
\title{Set and get a Flickr API key to/from environment variables.}
\usage{
setFlickrAPIKey(
  api_key,
  overwrite = FALSE,
  install = FALSE,
  call = caller_env()
)

set_flickr_api_key(
  api_key,
  overwrite = FALSE,
  install = FALSE,
  call = caller_env()
)

getFlickrAPIKey(api_key = NULL, strict = TRUE, call = caller_env())

get_flickr_api_key(api_key = NULL, strict = TRUE, call = caller_env())
}
\arguments{
\item{api_key}{Flickr API key. Optional for \code{\link[=getFlickrAPIKey]{getFlickrAPIKey()}}.}

\item{overwrite}{If \code{TRUE}, overwrite any existing API key in \code{.Renviron} using
the name "FLICKR_API_KEY". Defaults to \code{FALSE}.}

\item{install}{If \code{TRUE}, this function adds your API key to your \code{.Renviron}
using the name "FLICKR_API_KEY" for use in future sessions. Defaults to
\code{FALSE}.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}

\item{strict}{If \code{TRUE} (default), error if api_key is \code{NULL} and no valid
key is available using the "FLICKR_API_KEY" environment variable name. If
\code{FALSE}, warn instead of error and invisibly return \code{NULL}.}
}
\description{
\code{\link[=setFlickrAPIKey]{setFlickrAPIKey()}} sets a Flickr API key for the current session or for
future use as a \code{.Renviron} environment variable. \code{\link[=getFlickrAPIKey]{getFlickrAPIKey()}} gets a
saved Flickr API key from \code{.Renviron}.
}
