% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlickrAPIKey.R
\name{setFlickrAPIKey}
\alias{setFlickrAPIKey}
\alias{set_flickr_api_key}
\title{Set Flickr API key in environment variables}
\usage{
setFlickrAPIKey(api_key, overwrite = FALSE, install = FALSE)
}
\arguments{
\item{api_key}{Flickr API key}

\item{overwrite}{If `TRUE`, overwrite existing token; Default: FALSE}

\item{install}{If `TRUE`, install token for use in future sessions; Default: FALSE}
}
\description{
Set Flickr API key in environment variables
}
